// src/features/add_contract/model/contract.ts
export interface Contract {
  startDate: string;
  endDate: string;
  taxableSalary: number;
  salary: number;
  workDatePerWeek: number;
  note: string;
  noOfDependences: number;
  contractTypeId: number;
  salaryType: string;
  contractFile: string;
  createAt: string;
  responseId: number;
  changeAt: string;
  contractStatus: boolean;
}